456dev2/counted/Source/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.11",
    "pyjwt>=2.10.1",
    "uvicorn>=0.34.0",
]
